fixed the float

+++ src/eval_prm/run_eval_with_vllm.py
@@ -89,7 +89,7 @@ sampling_params = SamplingParams(
 )
 
 print(f"[INFO] Loading model: {args.model_name_or_path}")
-llm = LLM(model=args.model_name_or_path)
+llm = LLM(model=args.model_name_or_path, dtype="float16")
 
 # === Generate responses ===
 generations = []
